requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
qdrant-client
transformers
torch
sentence-transformers
easyocr
pytesseract
pdfplumber
pymupdf
sympy
accelerate
sentence-transformers
python-multipart
docx
llama-cpp-python